Create an AOI object

Create an object representing an Area of Interest (AOI). Only rectangles are supported (like a jpeg image in an experiment). Pixel (0,0) is the lower left corner of the screen.

Usage

create_aoi(aoi_name, x_pix, y_pix, screen_width = 1920, screen_height = 1080)

Arguments

aoi_name

label of the AOI

x_pix

location of the left and right edges in pixels.

y_pix

location of the bottom and top edges in pixels.

screen_width

width of the screen in pixels. Defaults to 1920.

screen_height

width of the screen in pixels. Defaults to 1080.

Value

an AOI object.