Create complementary response definitions

In the typical response definition, there is a primary response compared to other competitors. Oftentimes, we are interested in also comparing each of the competitors to the other images. This function quickly assembles a full cycle of response definitions.

cycle_response_def(response_def)

Arguments

response_def
a response definition to use a template for other definitions.

Value

a list of response definitions where each member of c(response_def$primary, response_def$others) is used as the primary response.

Examples

# Create one definition
def <- create_response_def(
  primary = 1,
  others = c(5, 8, 9),
  elsewhere = 0,
  missing = NA
)

# Create the full cycle of response definitions
cycle_response_def(def)

#> [[1]]
#> List of 5
#>  $ response_def: chr "1"
#>  $ primary     : num 1
#>  $ others      : num [1:3] 5 8 9
#>  $ elsewhere   : num 0
#>  $ missing     : logi NA
#>  - attr(*, "class")= chr "response_def"
#> 
#> [[2]]
#> List of 5
#>  $ response_def: chr "5"
#>  $ primary     : num 5
#>  $ others      : num [1:3] 1 8 9
#>  $ elsewhere   : num 0
#>  $ missing     : logi NA
#>  - attr(*, "class")= chr "response_def"
#> 
#> [[3]]
#> List of 5
#>  $ response_def: chr "8"
#>  $ primary     : num 8
#>  $ others      : num [1:3] 1 5 9
#>  $ elsewhere   : num 0
#>  $ missing     : logi NA
#>  - attr(*, "class")= chr "response_def"
#> 
#> [[4]]
#> List of 5
#>  $ response_def: chr "9"
#>  $ primary     : num 9
#>  $ others      : num [1:3] 1 5 8
#>  $ elsewhere   : num 0
#>  $ missing     : logi NA
#>  - attr(*, "class")= chr "response_def"
#>