Convert DataWiz data into long format

DataWiz files have several columns F0, F33, F67, etc. for each time sample. This function converts a dataframe from such a file into a long format, where there is a single time column and single column of gaze responses.

Usage

melt_datawiz(df, key_col = "Time", value_col = "Look")

Arguments

df

a dataframe created by reading a datawiz file

key_col

the name of the new column that holds the time values

value_col

the name of the new column that holds the looking data at each time sample

Value

a long data-frame